# openapi-contract-search.yaml
openapi: 3.0.3
info:
  title: SunnyBill Contract Search API
  description: API zum Suchen von Lieferantenverträgen anhand von Vertragskennungen
  version: 1.0.0
  contact:
    name: SunnyBill Support
servers:
  - url: http://localhost/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
  /contracts/search:
    get:
      summary: Suche Verträge anhand von Vertragskennungen
      description: |
        Sucht nach Lieferantenverträgen basierend auf den Vertragskennungen (contract_recognition_1, contract_recognition_2, contract_recognition_3).
        Gibt alle Verträge zurück, bei denen mindestens ein Key-Value-Paar übereinstimmt.
      operationId: searchContractsByIdentifiers
      tags:
        - Contracts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifiers
              properties:
                identifiers:
                  type: array
                  minItems: 1
                  description: Array von Key-Value-Paaren zum Durchsuchen der Vertragskennungen
                  items:
                    type: object
                    required:
                      - key
                      - value
                    properties:
                      key:
                        type: string
                        enum:
                          - contract_recognition_1
                          - contract_recognition_2
                          - contract_recognition_3
                        description: Feld, in dem gesucht werden soll
                      value:
                        type: string
                        description: Suchwert (unterstützt Teilstring-Suche)
                relations:
                  type: array
                  description: |
                    Optional - Welche Relationen sollen mitgeladen werden.
                    Unterstützt auch nested relations mit Punkt-Notation (z.B. "solarPlants.components").

                    Verfügbare Basis-Relationen:
                    - supplier
                    - solarPlants
                    - solarPlantAssignments
                    - billings
                    - contractNotes
                    - favoriteNotes
                    - standardNotes
                    - documents
                    - articles
                    - activeArticles
                    - contractMatchingRules
                    - activeContractMatchingRules

                    Verfügbare Nested-Relationen für solarPlants:
                    - solarPlants.components
                    - solarPlants.participations
                    - solarPlants.customers
                    - solarPlants.monthlyResults

                    Verfügbare Nested-Relationen für solarPlantAssignments:
                    - solarPlantAssignments.solarPlant
                    - solarPlantAssignments.solarPlant.components
                  items:
                    type: string
                    example: "solarPlants.components"
            examples:
              singleSearch:
                summary: Suche in einem Feld ohne Relations
                value:
                  identifiers:
                    - key: contract_recognition_1
                      value: "12345"
              multipleSearch:
                summary: Suche in mehreren Feldern mit Relations
                value:
                  identifiers:
                    - key: contract_recognition_1
                      value: "12345"
                    - key: contract_recognition_2
                      value: "ABCDE"
                    - key: contract_recognition_3
                      value: "XYZ-789"
                  relations:
                    - supplier
                    - billings
                    - documents
              allRelations:
                summary: Suche mit allen verfügbaren Relations
                value:
                  identifiers:
                    - key: contract_recognition_1
                      value: "12345"
                  relations:
                    - supplier
                    - solarPlants
                    - solarPlantAssignments
                    - billings
                    - contractNotes
                    - documents
                    - articles
              nestedRelations:
                summary: Suche mit nested Relations für Solaranlagen-Details
                value:
                  identifiers:
                    - key: contract_recognition_1
                      value: "12345"
                  relations:
                    - supplier
                    - solarPlants.components
                    - solarPlants.participations
                    - solarPlants.customers
                    - solarPlantAssignments.solarPlant
                    - billings
                    - documents
      responses:
        '200':
          description: Erfolgreiche Suche
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SupplierContract'
                  summary:
                    type: object
                    properties:
                      total_found:
                        type: integer
                        description: Anzahl der gefundenen Verträge
                        example: 5
                      search_criteria:
                        type: array
                        description: Verwendete Suchkriterien
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            value:
                              type: string
                      loaded_relations:
                        type: array
                        description: Geladene Relationen
                        items:
                          type: string
              examples:
                successResponse:
                  summary: Erfolgreiche Antwort mit gefundenen Verträgen
                  value:
                    success: true
                    data:
                      - id: "9d1a2b3c-4d5e-6f7g-8h9i-0j1k2l3m4n5o"
                        supplier_id: "8c1a2b3c-4d5e-6f7g-8h9i-0j1k2l3m4n5o"
                        contract_number: "LV-0001"
                        contract_recognition_1: "12345"
                        contract_recognition_2: "ABCDE"
                        contract_recognition_3: "XYZ-789"
                        title: "Energieliefervertrag"
                        status: "active"
                        supplier:
                          id: "8c1a2b3c-4d5e-6f7g-8h9i-0j1k2l3m4n5o"
                          company_name: "Energie AG"
                        billings: []
                        documents: []
                    summary:
                      total_found: 1
                      search_criteria:
                        - key: "contract_recognition_1"
                          value: "12345"
                      loaded_relations:
                        - supplier
                        - billings
                        - documents
        '422':
          description: Validierungsfehler
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Validierungsfehler"
                  errors:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
              examples:
                validationError:
                  summary: Validierungsfehler
                  value:
                    success: false
                    message: "Validierungsfehler"
                    errors:
                      identifiers:
                        - "The identifiers field is required."
        '401':
          description: Nicht authentifiziert
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthenticated."
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        App-Token für API-Zugriff. Benötigte Permission: suppliers:read

        Header-Format: Authorization: Bearer YOUR_APP_TOKEN
...